# zone/docs/swagger.yaml
definitions:
  httphandler.ShelterDTO:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      shelter_availability:
        type: boolean
      shelter_id:
        type: string
      shelter_name:
        type: string
      trail_id:
        type: string
    type: object
  httphandler.TrailDTO:
    properties:
      end_latitude:
        type: number
      end_longitude:
        type: number
      start_latitude:
        type: number
      start_longitude:
        type: number
      trail_id:
        type: string
      trail_name:
        type: string
      zone_id:
        type: string
    type: object
  httphandler.ZoneDTO:
    properties:
      zone_id:
        type: string
      zone_name:
        type: string
    type: object
info:
  contact:
    email: shil9@mcmaster.ca
    name: Liuyin Shi
    url: https://github.com/XIAOKAOBO
  description: This provides a description of API endpoints for the zone
  title: Zone Service API
  version: "1.0"
paths:
  /api/v1/zone:
    post:
      consumes:
      - application/json
      description: Create a new zone
      parameters:
      - description: Zone Data
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/httphandler.ZoneDTO'
      produces:
      - application/json
      responses:
        "201":
          description: "message\":\t\t\t\"zone created"
          schema:
            type: string
        "400":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a zone
      tags:
      - zone
  /api/v1/zone/{zone_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: error, message: success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a zone
      tags:
      - zone
    put:
      consumes:
      - application/json
      description: Update details of an existing zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Updated Zone Data
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/httphandler.ZoneDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: error, message: success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a zone
      tags:
      - zone
  /api/v1/zone/{zone_id}/trail:
    get:
      consumes:
      - application/json
      description: Get the closest trail based on given longitude and latitude in
        a specific zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success, message: closest trail, trail_id: UUID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: failed to retrieve closest trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the closest trail
      tags:
      - zone
    post:
      consumes:
      - application/json
      description: Create a trail based on given longitude and latitude within a zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Trail Data
        in: body
        name: trail
        required: true
        schema:
          $ref: '#/definitions/httphandler.TrailDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'status: success, message: trail created successfully, trail_id:
            UUID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: failed to create trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a trail
      tags:
      - zone
  /api/v1/zone/{zone_id}/trail/{trail_id}:
    delete:
      consumes:
      - application/json
      description: Delete a trail from the zone by its ID
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Trail ID
        in: path
        name: trail_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success, message: deleted trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: failed to delete trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a specific trail
      tags:
      - zone
    get:
      consumes:
      - application/json
      description: Retrieve detailed location information of a trail by its ID
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Trail ID
        in: path
        name: trail_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UUID of the closest trail
          schema:
            type: string
        "400":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Failed'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get location information of a specific trail
      tags:
      - zone
    put:
      consumes:
      - application/json
      description: Update details of an existing trail within a zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Trail ID
        in: path
        name: trail_id
        required: true
        type: string
      - description: Updated Trail Data
        in: body
        name: trail
        required: true
        schema:
          $ref: '#/definitions/httphandler.TrailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success, message: updated trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: failed to update trail'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a trail
      tags:
      - zone
  /api/v1/zone/{zone_id}/trail/{trail_id}/shelter:
    get:
      consumes:
      - application/json
      description: Retrieve the closest shelter information based on current longitude
        and latitude
      responses: {}
      summary: Get the closest shelter information
      tags:
      - zone
    post:
      consumes:
      - application/json
      description: Create a new shelter associated with a trail in a zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: string
      - description: Trail ID
        in: path
        name: trail_id
        required: true
        type: string
      - description: Shelter Data
        in: body
        name: shelter
        required: true
        schema:
          $ref: '#/definitions/httphandler.ShelterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'status: success, message: shelter created successfully, shelter_id:
            UUID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'status: error, message: failed to create shelter'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'status: error, message: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a shelter
      tags:
      - zone
...
swagger: "2.0"